Fixed issue with Queue entry allocation and deallocation
The Queue allocates memory using 'nextPowerOfTwo_(num_entries*2)'. For example,
if 5 entries are requested, storage is allocated for 16 entries. The queue was
not managing the construction/destruction of the Queue entries properly allowing
all 16 entries to concurrently be valid, constructed objects even though the
user only request 5.  Fixed this by using a placement new when pushing an entry
and by explicitly calling the destructor when invalidating an entry.