Add clang/clangxx explicitly to build deps
clang on linux is not fully independent.  The conda-forge guys build with clang
on linux by leaving gcc as the {{ compiler('c') }} and explicitly adding clang/clangxx
Because of this, they don't have a clang_linux-64 metapackage and things are a bit weird

Doing this workaround for now.  May want to create clang_linux-64 metapackage in case we
decide to build using both gcc and clang on linux.