disable Valgrind and use 'make regress' instead of raw ctest
Was trying to generate Test.xml using ctest but '--test-action test'
only runs the tests and our cmake setup doesn't have a 'build the tests
but don't run them' target.  Punting on that for now.

There also seems to be issues with valgrind.  The osx build is stale
compared to linux.  Going to look into that before enabling it.
see https://github.com/conda-forge/valgrind-feedstock/pull/9