Add PublishTestResults task to pipelines
Look for the cTest Testing/<datestamp>/Test.xml output files and
load them into the per-test results UI.  Will give us tracking of when
individual tests have pass/failed and whether they are flaky.

This needs to be a PR to conda-smithy to enable customized 'hook'
tasks to be inserted in the yaml it generates.  The next
'conda smithy rerender' will blow these changes away.