fix osx segfault on menu creation #113
If we are going to support more than Linux, we can't hardcode GTK+
icon names anywhere in the code.  Swap them out for existing WxWidgets
icons so that the segfaults stop on MacOS launching.

Will likely need to supply nicer looking icons in the resources folder
because some of these are really ugly i.e. wx.ART_TICK_MARK.