SyncPort recirculation assertion
If the SyncInPort is not ready, a packet that is already in flight, is recirculated, until the inPort becomes ready. In this case, the packet is to be delivered in the next cycle.
The problem with the current implementation is that it tries to recirculate the packet in the current cycle requiring a zero cycle latency port. This causes an assertion.
This commit adds 1 cycle latency preventing the assertion and in addition ensures, that the SyncInPort behaves as documented.