Removing parameter from isReady
The parameter for isReady to check the readiness of a connected SyncInPort can be omitted. In the documentation it is suggested, to check for  SyncOutPort::isReady(n), if this is followed by a SyncOutPort::send(message, n). However, between the current clock cycle `c` and `c+n`, the input port can become unready by calling SyncInPort::setReady(false). Therefore, there is a discrepancy of what the method `isReady` does and the comment. 
Further, `isReady` calls `couldAccept_` which then asserts that the parameter given to `isReady`, equals 0. 
This commit removes the parameter from the list of `isReady` and `couldAccept_`. Further, it improve comments regarding the purpose and functionality of `isReady`.