Fixing the "sliding" of packets to the next clock cycle
This commit fixes the "allow_slide" = true setting. This setting indicates, if packets are allowed to move to the next clock cycle, if packets are already in flight and scheduled to be transmitted in the same clock cycle. Since SyncPort does not allow the transmission of multiple packets within a clock cycle, setting allow_slide to true will automatically reschedule new packets to the next possible clock cycle.