skip tests on osx_arm64 in CI
CI cross-compiles osx_arm64 from a x86 build host and our scripts
don't support running the tests on an x86 host targeting arm.  It's
not worth the effort to add running the tests on that platform when
CI will switch to native builds in a few months anyway and this problem
will resolve itself.