DatabaseCheckpointer performance optimizations (#612)
I added a variety of utilities/APIs to SimDB to get `loadCheckpoint()`
numbers way down. These changes result in most checkpoints being
loadable from the cache directly (blazing fast) and now from the
pipeline task's queues (very fast). Similar optimizations are done for
checkpoint deletion.

When not found in the cache or pipeline, the fallback plan of just
querying the database can still take 1-8ms or so. Without these changes
in this PR, most use cases fell into that bucket. But retrieving
directly from the pipeline when we are able to (even with "undoing"
`boost::serialization` and zlib) takes around **20 microseconds** now to
perform `loadCheckpoint()`.

The CoSim event pipeline (500 microseconds) used to set the standard,
but is now the bottleneck =D