map_v2.1.10 (#627)
This PR redesigns the DatabaseCheckpointer since there are sporadic
failures seen in the CoSim test harness. There are more/less failures
depending on how you parameterize the harness (how many N steps to take
before flushing to N-1, how big of a cache to use, and how big of a
snapshot threshold to use).

The previous design fundamentally worked like this:
- LRU main cache, pipeline, DB
- Checkpoints can only exist in one place out of the above three
- Reloading a checkpoint from the pipeline will remove it from the
pipeline and add it back to the cache
- Reloading a checkpoint from the DB will delete the record from the DB
and add it back to the cache
- Reloading a checkpoint from anywhere deletes all future checkpoints
since it doesn't allow >1 branch
- Checkpoint IDs must be auto-incrementing and contiguous / no gaps 

The constant reloading / deleting future checkpoints is not a stable
design. All sporadic failures come down to exceptions from not being
able to find a given checkpoint ID. Making this a stable design would
require more thread contention / mutex locks and performance would not
be good. The reason there is a 100% pass rate with the default harness
parameters is that the pipeline is not very deep, and the cache is big
enough to avoid so many things being reloaded into the LRU cache and
deleted from the pipeline / DB. Change the parameters to deep pipelines
or a smaller cache, and all the bugs appear.

This new design is very different, but has many advantages:
- Full reuse of the FastCheckpointer to implement the cache. Multiple
branches allowed.
- Cache has no size limit and is only cleared with an explicit API call.
- All CoSim flush operations will be done from the cache guaranteed.
Much faster, no contention.
- When you are ready to move the simulation forward, pick one of the
branches / checkpoints from the FastCheckpointer, and let those be
committed down the pipeline. All checkpoints not committed to the
pipeline will just be removed from the FastCheckpointer.
- Think of the cache as a staging area for git workflows. Create
branches, commit changes, delete unused branches, move on.
- Use non-contiguous checkpoint IDs (which contain those from branches
NOT taken) and arch IDs (these are now the contiguous IDs only from
checkpoints / branches taken and sent to the pipeline).
- Limitation: Once you commit checkpoints (remove from cache), you
cannot flush back to those checkpoints. But you are in control of how
big the cache gets before clearing it with the API call. We could relax
this limitation in the future, but for CoSim we will never have a need
to flush to an older event that was already committed. So this is not a
limitation for Pegasus.

I am wrapping up the follow-up PR for Pegasus which integrates this
checkpointer. Previously, the pass rate could be anywhere from 30% to
85% for some harness configurations, and the CoSim pass rate is now 100%
for all harness configurations I've tried.

One last note: I can follow up with another PR that renames
CherryPickFastCheckpointer back to DatabaseCheckpointer, but I did it
this way or the PR diff would be too hard to read since everything
changed. But there is also the PersistentFastCheckpointer, which sounds
the same as DatabaseCheckpointer, so I thought having
CherryPickFastCheckpointer makes sense since it's the only one that has
this behavior of pulling checkpoints / branches out of the
FastCheckpointer.