Bump SimDB (#633)
Need pipeline snoopers for CoSim event recreation performance.

Note that this was already a feature in SimDB, then the pipelines got
redesigned. I'm adding it back for CoSim which currently uses no
snoopers. When looking for something not in the event cache, it
currently needs to flush the whole pipeline to the database and query /
recreate the event from disk (slow). Snooping the pipeline stages in
memory is much faster, unless the requested event is so old that it is
in the database already. For most use cases, we should be able to get
the event from the boost serialization or zlib stages without flushing
or going to disk.

Next I'll release map_v2.1.12 which should be the last release we need
for Pegasus' basic cosim functionality.